requirements.txt
fastapi
uvicorn
numpy
faiss-cpu
httpx
python-dotenv
pydantic
